Переименование модуля
